refactor: renamed pages as routes
